agregue dependencia de email y configuracion de envio de email tambien valide que a ciertas rutas solo el admin pueda utilizarlas

+++ backend/explorer/src/main/java/com/example/explorer/controller/UserController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.explorer.service.UserService;
 import com.example.explorer.DTO.responseDTO;
+import com.example.explorer.model.User;
 
 @RestController
 @RequestMapping("/api/v1/user")
@@ -24,24 +26,30 @@ public class UserController {
 
     @GetMapping("/") // solo administrador
     public ResponseEntity<Object> getAllUsers() {
-        return new ResponseEntity<>(userService.findAll(), HttpStatus.OK);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (user.getRole().getId() != 2) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Acceso denegado. Solo administradores.");
+        }
+
+        return ResponseEntity.ok(userService.findAll());
     }
 
     @GetMapping("/profile")
     public ResponseEntity<Object> profile(@AuthenticationPrincipal UserDetails userDetails) {
-
         return new ResponseEntity<Object>(userDetails, HttpStatus.OK);
     }
 
-    @DeleteMapping("/{id}") // solo administrador, falta desarrollar
+    @DeleteMapping("/{id}") // solo administrador
     public ResponseEntity<Object> deleteUser(@PathVariable int id) {
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (user.getRole().getId() != 2) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Acceso denegado. Solo administradores.");
+        }
+
         responseDTO response = userService.deleteUser(id);
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    // @PostMapping("profile/")
-    // public ResponseEntity<String> profile(@RequestBody String request) {
-    // return new ResponseEntity<String>("end-point private profile",
-    // HttpStatus.OK);
-    // }
 }
